ADDS: begins working on data revalidation
checks the diff between current and incoming data (two methods, one (bad)
builds keys via json.stringify the entire structure to check diff and
the other (good) uses the updatedAt field to construct the key against
the docId to check diff with current data. Currently there is an
interval to reinvalidate the data, but because this gets tricky when
animations are happening, it will be better to do this invalidation
when, before, or after maximum announcements take over the screen, that
way there is no visual glitch when the data updates.