ADDS: authentication
Uses a not-authenticated screen which displays the device's registered
name (as see in the LRCo Admin Console), and provides instructions on
how to verify the device. The data is stored in the session storage. Not
sure how this persists, so may need to find a more long term solution as
this device should not need to be re-authenticated often.

All authenticated routes to api have been restored.