FIX: issues with token
The server sends back the verified token, which does not cross check the
verification of the device in the database, so I make an additional call
to post the token before asking for the authstate to update the token
and it now works. There is an unfortunate lag which flashes the not-auth
screen that I will look into later.