ADDS: improved sorting, diminishing page duration, ...
Now sorts sessions based on their start distance from now, so sessions
that have just started, or are just about to start are shown on the
first page, and sessions which have been going for a while or start a
while from **now** are shown on later pages.

To further this improvement the session pages now feature a diminishing
duration, where the first page (featuring the most important
information) run the longest, and the subsuquent pages diminish their
duration, down to a minimum of 6 seconds for pages 4+ with the least
important information.