HACKFIX: kinda fixed the animation degration issue
Fix max ann runs perfect, but all subsequent transitions occur before
progress bar completes, so I altered the duration for the progress bar
to 1200 (1000 for conversion from ms to s, then 200 for the degration).

Now need to check that other uses of the progress bar are not also
broken with this fix. If so I'll need to find the issue in MaxAnns.tsx