HACK/FIX: removed previous fix. This is more direct to maxanns
The previous fix had issues with other uses of ProgressBar. MaxAnns is
the only instance of the progess bar where is issue occurs. So,
calculated the animation_offset to subtract from the progress bar time.
At an arbitrary value I got it to feel correct.

The strangest part about this is that the first announcement runs
perfectly. Logs show that the animation duration is the same as the
bar's duration. However, every subsequent animation terminates before
the progess bar at a CONSTANT RATE.